Simplify the model, use javaVersion to store the target level
