+++ plugin/src/test/java/com/microsoft/java/bs/gradle/plugin/GradleBuildServerPluginTest.java
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 
 import com.microsoft.java.bs.gradle.model.GradleSourceSet;
 import com.microsoft.java.bs.gradle.model.GradleSourceSets;
-import com.microsoft.java.bs.gradle.model.JdkPlatform;
 
 class GradleBuildServerPluginTest {
 
@@ -56,11 +55,8 @@ class GradleBuildServerPluginTest {
         assertNotNull(gradleSourceSet.getSourceOutputDir());
         assertNotNull(gradleSourceSet.getResourceOutputDir());
 
-        JdkPlatform jdkPlatform = gradleSourceSet.getJdkPlatform();
-        assertNotNull(jdkPlatform.getJavaHome());
-        assertNotNull(jdkPlatform.getJavaVersion());
-        assertNotNull(jdkPlatform.getSourceCompatibility());
-        assertNotNull(jdkPlatform.getTargetCompatibility());
+        assertNotNull(gradleSourceSet.getJavaHome());
+        assertNotNull(gradleSourceSet.getJavaVersion());
       }
     }
